feat: update worker with export error
